% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{pred_score}
\alias{pred_score}
\title{Function to predict the sore of a football match.}
\usage{
pred_score(result, ratio, max_add_goals = 2)
}
\arguments{
\item{result}{A string. 'home', 'draw' or 'away'}

\item{ratio}{The probability that the home team wins}

\item{max_add_goals}{Maximum additional goals if the home team wins}
}
\value{
A string of the result
}
\description{
Function to predict the sore of a football match.
}
\examples{
pred_score("home", 0.8)
}
