% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{prob_a_beats_b}
\alias{prob_a_beats_b}
\title{Calculates tthe probability, that team a beats team b.
For both teams, there have to be a measurement of numeric
measurement of their strength. For example an elo number.}
\usage{
prob_a_beats_b(elo_team_a, elo_team_b)
}
\arguments{
\item{elo_team_a}{Elo team a}

\item{elo_team_b}{Elo team b}
}
\value{
Probability, that team a beats team b
}
\description{
Calculates tthe probability, that team a beats team b.
For both teams, there have to be a measurement of numeric
measurement of their strength. For example an elo number.
}
