% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fixtures.R
\name{read_competitions}
\alias{read_competitions}
\title{Function to read informations about
all competitions available}
\usage{
read_competitions(token)
}
\arguments{
\item{token}{}
}
\description{
Function to read informations about
all competitions available
}
