% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fixtures.R
\name{read_fixtures}
\alias{read_fixtures}
\title{Return all fixtures of a certain league.}
\usage{
read_fixtures(token, com_id)
}
\arguments{
\item{league}{}
}
\description{
Return all fixtures of a certain league.
}
