% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fixtures.R
\name{read_team_info}
\alias{read_team_info}
\title{Read information about every team in a certain competition}
\usage{
read_team_info(token, com_id)
}
\arguments{
\item{com_id}{}
}
\description{
Read information about every team in a certain competition
}
